+++ packages/react-components/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @lualtek/react-components
 
+## 15.0.0-next.0
+
+### Major Changes
+
+- 7b45869: Migrate theme system to css light-dark()
+
+### Minor Changes
+
+- 5612bd4: Update deps
+
+### Patch Changes
+
+- ecb5ae0: Fix shandow lv 3
+- f9de811: Update deps
+- Updated dependencies [5612bd4]
+  - @lualtek/icons@15.0.0-next.0
+
 ## 14.4.0
 
 ### Minor Changes